+++ galaxy_api/api/pagination.py
@@ -17,11 +17,6 @@
 
 import logging
 
-from collections import OrderedDict
-
-from drf_yasg import openapi
-from drf_yasg.inspectors import PaginatorInspector
-
 from rest_framework import pagination
 from rest_framework import response
 from rest_framework.utils.urls import replace_query_param
@@ -100,51 +95,3 @@ class InsightsStylePagination(pagination.LimitOffsetPagination):
             },
             'data': data
         })
-
-class IPP12RestResponsePagination(PaginatorInspector):
-    """
-    Provides the response schema to match the output of
-    CustomPageNumberPagination as per IPP-12.
-    """
-    def get_paginated_response(self, paginator, response_schema):
-        if not isinstance(paginator, InsightsStylePagination):
-            return super(PaginatorInspector, self).get_paginated_response(paginator, response_schema)
-        return openapi.Schema(
-            type=openapi.TYPE_OBJECT,
-            properties=OrderedDict((
-                ('meta', openapi.Schema(
-                    type=openapi.TYPE_OBJECT,
-                    properties=OrderedDict((
-                        ('count', openapi.Schema(
-                            type=openapi.TYPE_INTEGER,
-                        )),
-                    )),
-                    required=['count'],
-                )),
-                ('links', openapi.Schema(
-                    type=openapi.TYPE_OBJECT,
-                    properties=OrderedDict((
-                        ('first', openapi.Schema(
-                            type=openapi.TYPE_STRING, format=openapi.FORMAT_URI,
-                            x_nullable=True
-                        )),
-                        ('previous', openapi.Schema(
-                            type=openapi.TYPE_STRING, format=openapi.FORMAT_URI,
-                            x_nullable=True
-                        )),
-                        ('next', openapi.Schema(
-                            type=openapi.TYPE_STRING, format=openapi.FORMAT_URI,
-                            x_nullable=True
-                        )),
-                        ('last', openapi.Schema(
-                            type=openapi.TYPE_STRING, format=openapi.FORMAT_URI,
-                            x_nullable=True
-                        )),
-                    )),
-                )),
-                ('data', response_schema),
-            )),
-            required=['data'],
-        )
-
-
